api/requirements.txt
django
djangorestframework
django-cors-headers
drf-spectacular
psycopg2
uwsgi
